% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matterArray.R
\docType{package}
\name{matterArray-package}
\alias{matterArray-package}
\title{matter back end for DelayedArray objects.}
\description{
An array-like container for convenient access and manipulation of matter
datasets. Supports delayed operations and block processing.
}
